Rename BingoTracker and certain methods for clarity
