+++ BingoSync/BingoSync.csproj
@@ -13,8 +13,8 @@
 		<Product>BingoSync</Product>
 		<Description>A Hollow Knight Mod</Description>
 		<Copyright>Copyright ©  2024</Copyright>
-		<AssemblyVersion>1.4.1.0</AssemblyVersion>
-		<FileVersion>1.4.1.0</FileVersion>
+		<AssemblyVersion>1.4.1.1</AssemblyVersion>
+		<FileVersion>1.4.1.1</FileVersion>
 		<OutputPath>bin\$(Configuration)\</OutputPath>
 		<LangVersion>latest</LangVersion>
 		<GenerateDocumentationFile>True</GenerateDocumentationFile>
