+++ BingoSync/CustomVariables/GiantGeoEgg.cs
@@ -16,7 +16,7 @@ namespace BingoSync.CustomVariables
             bool hasDestroyedState = self.TryGetState(destroyedStateName, out FsmState destroyedState);
             if (self == null || self.FsmName != fsmName || !hasDestroyedState) return;
             if (self.gameObject == null || self.gameObject.name != objectName) return;
-            destroyedState.AddCustomAction(() => BingoTracker.UpdateBoolean(variableName, true));
+            destroyedState.AddCustomAction(() => GoalCompletionTracker.UpdateBoolean(variableName, true));
         }
     }
 }
