+++ BingoSync/CustomVariables/Jiji.cs
@@ -14,7 +14,7 @@ namespace BingoSync.CustomVariables
             if (zone != MapZone.TOWN)
                 return orig;
 
-            BingoTracker.UpdateBoolean(variableName, true);
+            GoalCompletionTracker.UpdateBoolean(variableName, true);
             return orig;
         }
     }
