+++ BingoSync/CustomVariables/Telescope.cs
@@ -17,7 +17,7 @@ namespace BingoSync.CustomVariables
             if (self == null || self.FsmName != fsmName || !hasInteractState) return;
             if (self.gameObject == null || self.gameObject.name != objectName) return;
             interactState.AddCustomAction(() => {
-                BingoTracker.UpdateBoolean(variableName, true);
+                GoalCompletionTracker.UpdateBoolean(variableName, true);
             });
         }
     }
