+++ BingoSync/Hooks.cs
@@ -159,19 +159,19 @@ namespace BingoSync
 
         private static void HeroUpdate()
         {
-            BingoTracker.ProcessBingo(Controller.ActiveSession);
+            GoalCompletionTracker.UpdateAllKnownSquares(Controller.ActiveSession);
         }
 
         private static bool UpdateBoolInternal(string name, bool orig)
         {
-            BingoTracker.UpdateBoolean(name, orig);
+            GoalCompletionTracker.UpdateBoolean(name, orig);
             return orig;
         }
 
         private static int UpdateIntInternal(string name, int current)
         {
             var previous = PlayerData.instance.GetIntInternal(name);
-            BingoTracker.UpdateInteger(name, previous, current);
+            GoalCompletionTracker.UpdateInteger(name, previous, current);
             return current;
         }
